% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{boundsKendallStar}
\alias{boundsKendallStar}
\title{boundsKendallStar}
\usage{
boundsKendallStar(x, y, W)
}
\arguments{
\item{x}{data vector}

\item{y}{data vector}

\item{W}{weight matrix, number of columns correspond to length of x and y.}
}
\value{
\code{list} list with two objects named "upper" and "lower", vectors of upper and lower bounds on the association measure
}
\description{
The boundsKendallStar function calculates the upper and lower bounds the weighted zero-inflated Kendall's tau star association measure, given a matrix containing a number of weights.
}
\examples{

x = pmax(0,rnorm(100))
y = pmax(0,rnorm(100))
w = runif(100)
wcor = weightedKendallStar(x,y,w)
boundsKendallStar(x,y,w)

}
