% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{DCARS}
\alias{DCARS}
\title{DCARS}
\usage{
DCARS(dat, xname, yname, W = NULL, rangeMin = 0, wcormin = 0,
  statmin = 0, weightedConcordanceFunction = weightedPearson,
  weightedConcordanceFunctionW = "vector",
  extractTestStatisticOnly = FALSE, extractWcorSequenceOnly = FALSE,
  plot = FALSE, niter = 100,
  extractPermutationTestStatistics = FALSE, verbose = FALSE,
  forceBlock = FALSE, WcorSequenceSummaryFun = sd, ...)
}
\arguments{
\item{dat}{a genes x samples gene expression rank matrix, should be already converted to ranks with first column lowest survival and last column highest survival}

\item{xname}{name of row of dat to test together with yname}

\item{yname}{name of row of dat to test together with xname}

\item{W}{weight matrix for weighted correlations,}

\item{rangeMin}{minimum range of weighted correlation vector to include for permutation testing}

\item{wcormin}{minimum absolute value weighted correlation vector to include for permutation testing}

\item{statmin}{minimum value DCARS test statistic to include for permutation testing}

\item{weightedConcordanceFunction}{concordance function to use, defaults to weighted Pearson correlation. User can provide their own function, with arguments fun(x,y,w).}

\item{weightedConcordanceFunctionW}{either "vector" or "matrix", determining if the function given in weightedConcordanceFunction argument takes in a single vector as input for w, or a matrix.}

\item{extractTestStatisticOnly}{if TRUE, extract only the DCARS test statistic without permutation testing}

\item{plot}{if TRUE plot observed weighted correlatin vector}

\item{niter}{number of iterations for permutation testing}

\item{extractPermutationTestStatistics}{if TRUE, extract only the DCARS test statistic without permutation testing}

\item{verbose}{if TRUE, print updates}

\item{forceBlock}{if TRUE, this converts all positive values in weight matrix W to 1, and calculates weighted correlation on the resulting subset. Default FALSE}

\item{WcorSequenceSummaryFun}{Function that defines how the local weighted correlations are summarised, default is the function sd for standard deviation.}

\item{...}{additional arguments passing on to weightMatrix()}

\item{extractWcorSequence}{if TRUE, extract only the weighted correlation vector without permutation testing}
}
\value{
either single value (p-value or test statistic), vector (local weighted correlation), or list (combination of above) depending on the input parameters.
}
\description{
The DCARS function performs testing for differential correlation across ranked samples.
}
\examples{

data(STRING)
data(SKCM)
SKCM_rank = t(apply(SKCM,1,rank))

# highly significantly DCARS gene pair: SKP1 and SKP2
# calculates p-value based on permutation
DCARS(SKCM_rank,"SKP1","SKP2",plot=TRUE)
# extract only the test statistic
DCARS(SKCM_rank,"SKP1","SKP2", extractTestStatisticOnly = TRUE)

# not significantly DCARS gene pair: EIF3C and EIF5B
# calculates p-value based on permutation
DCARS(SKCM_rank,"EIF3C","EIF5B",plot=TRUE)
# extract only the test statistic
DCARS(SKCM_rank,"EIF3C","EIF5B", extractTestStatisticOnly = TRUE)

# build weight matrix
W = weightMatrix(ncol(SKCM_rank), type = "triangular", span = 0.5, plot = TRUE)

# extract DCARS test statistics
SKCM_stats = DCARSacrossNetwork(SKCM_rank,edgelist = STRING,
                                W = W, extractTestStatisticOnly = TRUE,
                                verbose = FALSE)
sort(SKCM_stats,decreasing=TRUE)[1:10]

}
