% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{DCARSacrossNetwork}
\alias{DCARSacrossNetwork}
\title{DCARSacrossNetwork}
\usage{
DCARSacrossNetwork(dat, edgelist, edgeNames = rownames(edgelist), ...)
}
\arguments{
\item{dat}{a genes x samples gene expression rank matrix, should be already converted to ranks with first column lowest survival and last column highest survival}

\item{edgelist}{is a 2 column character matrix with the genes to test for DCARS. if edgelist has more than 2 columns then the first two are taken}

\item{edgeNames}{is the name to assign to each edge, defaults to rownames(edgelist). if edgelist has no rownames then defaults to two gene names pasted with "_" in alphabetical order}

\item{...}{additional parameters passed to DCARS()}
}
\value{
value of this function depends on arguments passed into the DCARS() function (e.g. if extractTestStatisticOnly is set as TRUE)
}
\description{
performs DCARS method across all edges listed in the network for the (already ranked) genes x samples matrix dat
}
\examples{
data(STRING)
data(SKCM)
SKCM_rank = t(apply(SKCM,1,rank))

# highly significantly DCARS gene pair: SKP1 and SKP2
# calculates p-value based on permutation
DCARS(SKCM_rank,"SKP1","SKP2",plot=TRUE)
# extract only the test statistic
DCARS(SKCM_rank,"SKP1","SKP2", extractTestStatisticOnly = TRUE)

# not significantly DCARS gene pair: EIF3C and EIF5B
# calculates p-value based on permutation
DCARS(SKCM_rank,"EIF3C","EIF5B",plot=TRUE)
# extract only the test statistic
DCARS(SKCM_rank,"EIF3C","EIF5B", extractTestStatisticOnly = TRUE)

# build weight matrix
W = weightMatrix(ncol(SKCM_rank), type = "triangular", span = 0.5, plot = TRUE)

# extract DCARS test statistics
SKCM_stats = DCARSacrossNetwork(SKCM_rank,edgelist = STRING,
                                W = W, extractTestStatisticOnly = TRUE,
                                verbose = FALSE)
sort(SKCM_stats,decreasing=TRUE)[1:10]
}
