% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{estimatePvaluesSpearman}
\alias{estimatePvaluesSpearman}
\title{estimatePvaluesSpearman}
\usage{
estimatePvaluesSpearman(stats, globalCors, permstats, usenperm = FALSE,
  nperm = 10000, maxDist = 0.1, plot = FALSE, verbose = FALSE)
}
\arguments{
\item{stats}{a named vector of observed DCARS test statistics, names should appear in names of globalCor}

\item{globalCors}{a named vector containing global spearman correlation for each gene pair}

\item{permstats}{a named list containing permuted test statistics, names should appear in names of globalCor}

\item{usenperm}{logical to use number of permutations to select nearest genepairs}

\item{nperm}{number of permutations, used is usenperm = TRUE}

\item{maxDist}{numeric maximum difference in globalCor to consider for permutations, only used if usenperm = FALSE}

\item{plot}{logical default FALSE, if TRUE plots the globalCor against mean of permuted test statistics, and shows fitted polynomial}

\item{verbose}{logical default FALSE, if TRUE prints genepair names}
}
\value{
\code{data.frame}
}
\description{
the estimatePvaluesSpearman function more accurately estimates p-values by taking into account the relationship between the global correlation and the distribution of the permuted Spearman-based DCARS test statistics.
}
\examples{


}
