% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{fisherZtransformTest}
\alias{fisherZtransformTest}
\title{fisherZtransformTest}
\usage{
fisherZtransformTest(dat, xname, yname, classLabels = NULL)
}
\arguments{
\item{dat}{dat is a gene expression matrix (rows genes columns samples)}

\item{xname}{the name of the first gene}

\item{yname}{name of the second gene}

\item{classLabels}{a vector of two class labels, if NULL then split into 50/50. odd number of samples will ignore the middle sample}
}
\value{
single value for p-value
}
\description{
Performs Fisher's Z Transformation test for differential correlation.
}
\examples{

data(STRING)
data(SKCM)
fisherZtransformTest(dat = SKCM,xname = "EIF3C",yname = "EIF5B")

}
