% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{getLoessCriticalValue}
\alias{getLoessCriticalValue}
\title{getLoessCriticalValue}
\usage{
getLoessCriticalValue(stats, pvals, signifValue = 0.05, plot = FALSE)
}
\arguments{
\item{stats}{a vector of DCARS test statistics for which permutation testing has been done}

\item{pvals}{a vector of pvals for DCARS test}

\item{signifValue}{significance value (default 0.05)}

\item{plot}{logical (default FALSE)}
}
\value{
\code{numeric} a critical value for the unadjusted significance value
}
\description{
the getLoessCriticalValue function
}
\examples{

stats = rchisq(100,1)
pvals = 1 - pchisq(stats,1)
criticalValue = getLoessCriticalValue(stats, pvals, plot = TRUE)
criticalValue

}
