% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{LinearModelInteractionTest}
\alias{LinearModelInteractionTest}
\title{LinearModelInteractionTest}
\usage{
LinearModelInteractionTest(dat, xname, yname, response = NULL)
}
\arguments{
\item{dat}{dat is a gene expression matrix (rows genes columns samples)}

\item{xname}{the name of the first gene}

\item{yname}{name of the second gene}

\item{response}{a vector of two class labels, if NULL then split into 50/50. odd number of samples will ignore the middle sample}
}
\value{
single value for p-value
}
\description{
Fits a linear model with interaction term and tests for significance of the interaction term.
}
\examples{

data(STRING)
data(SKCM)
LinearModelInteractionTest(dat = SKCM,xname = "EIF3C",yname = "EIF5B")

}
