% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{logistic}
\alias{logistic}
\title{logistic}
\usage{
logistic(p)
}
\arguments{
\item{x}{numeric vector of non-negative data x}

\item{y}{numeric vector of non-negative data y}

\item{offset}{should 1 be added when pospos is 0 or all}
}
\value{
\code{numeric} of weighted correlations for the sequence of weights given
}
\description{
The logistic function calculates the (inverse) logistic for a given p, where p lies between 0 and 1. Large values indicates many zeros.
}
\examples{

p = seq(0.1,0.9, length.out = 100)
l = logistic(p)
plot(p,l)

}
