% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{plotColouredExpression}
\alias{plotColouredExpression}
\title{plotColouredExpression}
\usage{
plotColouredExpression(branchData, genepair, subsetBranch = NULL,
  n = 3, fittedline = TRUE)
}
\arguments{
\item{branchData}{is a list containing matrices of the cell expression per branch, assumed that the columns of each matrix in branchData is ordered by pseudotime. If branchData is not given as a list, it will be converted into a list containing branchData.}

\item{genepair}{is either a single character string with an underscore, or a two length character vector}

\item{subsetBranch}{subsetBranch is a character vector containing the names of the branches to be plotted. If NULL it will plot all branches}

\item{n}{number of panels to split ranked samples into, default 3.}

\item{fittedline}{logical default TRUE, add a lm straight line to the plot}
}
\value{
\code{ggplot} a ggplot object of scatterplots of expression split by sample ordering
}
\description{
the plotColouredExpression function plots an n-panel scatterplot of the gene pairs split by early, mid, and late in the sample ordering.
}
\examples{






}
