% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{plotEgoNetwork}
\alias{plotEgoNetwork}
\title{plotEgoNetwork}
\usage{
plotEgoNetwork(hubnode, network, weight = "weight", subset = FALSE,
  thresh = NULL)
}
\arguments{
\item{hubnode}{is a character vector of node(s) to include as hub nodes}

\item{subset}{is a logical asking if you should subset based on the weight (default FALSE)}

\item{thresh}{is the subset weight threshold}

\item{g}{is an igraph network, with E(g)[[weight]] given as DCARS test statistics}
}
\value{
\code{igraph} object containing the network graphed. Produces an igraph plot
}
\description{
the plotEgoNetwork function plots network graphs with edges coloured by weights in the network
}
\examples{






}
