% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{plotNetworkPathway}
\alias{plotNetworkPathway}
\title{plotNetworkPathway}
\usage{
plotNetworkPathway(sigPairsList, minCommunity = 10, pathways = NULL,
  geneUniverse = NULL)
}
\arguments{
\item{sigPairsList}{list of significant pairs for branches, in igraph network form}

\item{minCommunity}{minimum number of genes in community to annotate most represented}

\item{pathways}{is a named list containing gene sets for querying, if left NULL defaults to REACTOME c2.all.v6.1.symbols.gmt downloaded from MSigDB}

\item{geneUniverse}{is the universe of genes measured, if null its taken as all the sigpairs}
}
\value{
\code{plot} plots of igraph objects
}
\description{
the plotNetworkPathway function plots network graphs with most represented pathways labelled
}
\examples{




}
