% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{plotOrderedExpression}
\alias{plotOrderedExpression}
\title{plotOrderedExpression plots expression vectors along branches and genes as ribbon plots}
\usage{
plotOrderedExpression(branchData, gene, xvals = NULL,
  subsetBranch = NULL, facet = FALSE)
}
\arguments{
\item{branchData}{is a list containing matrices of the cell expression per branch, assumed that the columns of each matrix in branchData is ordered by pseudotime. If branchData is not given as a list, it will be converted into a list containing branchData.}

\item{gene}{is either a single character string with an underscore, or a two length characer vector}

\item{xvals}{is a list containing the x-values associated with the samples in branchData (if NULL, samples will just be plotted against their rank)}

\item{subsetBranch}{subsetBranch is a character vector containing the names of the branches to be plotted. If NULL it will plot all branches}

\item{facet}{can either be FALSE, "branch", "gene", or "both"}
}
\value{
\code{ggplot} a ggplot object for ribbon plot with points
}
\description{
the plotOrderedExpression function
}
\examples{




}
