% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{plotWCorLine}
\alias{plotWCorLine}
\title{plotWCorLine plots weighted correlation vectors as line plots}
\usage{
plotWCorLine(wcorsList, gene)
}
\arguments{
\item{wcorsList}{is a list of matrices, with each matrix gene pair x samples weighted correlation vectors, assumed that they have same number of rows}

\item{gene}{is either a logical vector matching rows of entries in wcorsList, or a character of a gene}
}
\value{
\code{ggplot} object with line plots
}
\description{
the plotWCorLine function
}
\examples{






}
