% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_wqDCARS.R
\name{predict_DCAsignature}
\alias{predict_DCAsignature}
\title{Calculate DCAsignature}
\usage{
predict_DCAsignature(my_expr)
}
\arguments{
\item{my_expr}{A numeric matrix with rows as genes and columns as samples.}
}
\value{
A data frame with sample names and their DCAsignature scores.
}
\description{
This function calculates the DCAsignature based on a given gene expression matrix.
The expression matrix should have rows as genes and columns as samples.
If the matrix is missing required genes, an error will be thrown.
}
