% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{stratifiedSample}
\alias{stratifiedSample}
\title{stratifiedSample}
\usage{
stratifiedSample(stats, length = 100)
}
\arguments{
\item{stats}{a vector of statistics}

\item{length}{rough number of indices to extract}
}
\value{
\code{vector} a vector of indices for the statistics that should be used for permutation testing
}
\description{
the stratifiedSample function takes in a vector of statistics, and returns a vector of indices for the statistics that should be used for permutation testing
}
\examples{

stats = rnorm(5000)
sampleindices = stratifiedSample(stats)
length(sampleindices)

}
