% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{stretch}
\alias{stretch}
\title{stretch}
\usage{
stretch(wcor, upper = 1, lower = -1)
}
\arguments{
\item{wcor}{weighted correlation vector}

\item{upper}{either single numeric or vector of same length as wcor for upper bound of correlation. Default 1}

\item{lower}{either single numeric or vector of same length as wcor for lower bound of correlation. Default -1}
}
\value{
\code{vector} standardised weighted correlation vector
}
\description{
The stretch function calculates a 'stretched'/standardised vector of weighted correlations, using previously calculated upper and lower bounds of the statistic (same length vectors). Values are stretched higher or lower symmetrically around 0.
}
\examples{

x = pmax(0,rnorm(100))
y = pmax(0,rnorm(100))
w = runif(100)
wcor = weightedKendallStar(x,y,w)
stretch(wcor, upper = 0.8, lower = -0.3)

}
