% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{sweightedKendallStar}
\alias{sweightedKendallStar}
\title{sweightedKendallStar}
\usage{
sweightedKendallStar(x, y, w = 1)
}
\arguments{
\item{x}{data vector}

\item{y}{data vector}

\item{w}{weights vector}
}
\value{
\code{vector} of stretched weighted correlations using zero-inflated kendall's tau association measure
}
\description{
The sweightedKendallStar function is a convenience function to get swcor for weighted zero-inflated kendall's tau
}
\examples{

x = pmax(0,rnorm(100))
y = pmax(0,rnorm(100))
w = runif(100)
wcor = weightedKendallStar(x,y,w)
bounds = boundsKendallStar(x,y,w)
stretch(wcor, upper = bounds[["upper"]], lower = bounds[["lower"]]

sweightedKendallStar(x,y,w)

}
