% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{thin}
\alias{thin}
\title{thin}
\usage{
thin(W, n = 100)
}
\arguments{
\item{W}{matrix}

\item{n}{rough number of rows to keep}
}
\value{
\code{matrix} of thinned matrix keeping only roughly n rows.
}
\description{
The thin function extracts the rows of a matrix evenly so that roughly n number of rows remain. Used for thinning down the weight matrix to speed up overall computation.
}
\examples{

W = weightMatrix(500)
W_small = thin(W, n = 100)

}
