% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{weightedcor}
\alias{weightedcor}
\title{weightedcor}
\usage{
weightedcor(x, y, w)
}
\arguments{
\item{x}{x and y are data vectors}

\item{y}{x and y are data vectors}

\item{w}{weight vector}
}
\value{
weighted correlation value between x and y
}
\description{
Calculates weighted Pearson correlation between x and y.
This function has been superceded by weightedPearson() and
similar matrix function weightedPearson_matrix()
}
\examples{

x = rnorm(100)
y = rnorm(100)
w = runif(100)
weightedcor(x,y,w)

}
