% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{weightedPearson_matrix}
\alias{weightedPearson_matrix}
\title{weightedPearson_matrix}
\usage{
weightedPearson_matrix(x, y, W)
}
\arguments{
\item{x}{x and y are data vectors}

\item{y}{x and y are data vectors}

\item{W}{weight matrix, values should be between 0 and 1, number of columns should be the same as length(x) and length(y)}
}
\value{
\code{vector} weighted correlation values between x and y
}
\description{
The weightedPearson_matrix function calculates a vector of weighted correlations for two given data vectors, for a matrix of given weights.
}
\examples{

x = rnorm(100)
y = rnorm(100)
W = weightMatrix(100)
weightedPearson_matrix(x,y,w)

}
