% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{weightedSpearman}
\alias{weightedSpearman}
\title{weightedSpearman}
\usage{
weightedSpearman(x, y, w)
}
\arguments{
\item{x}{x and y are data vectors}

\item{y}{x and y are data vectors}

\item{w}{weight vector, values should be between 0 and 1}
}
\value{
\code{numeric} weighted correlation value between x and y
}
\description{
the weightedSpearman function
}
\examples{

x = rnorm(100)
y = rnorm(100)
w = runif(100)
weightedSpearman(x,y,w)

}
