% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{weightedStretchedKendallStar_matrix}
\alias{weightedStretchedKendallStar_matrix}
\title{weightedStretchedKendallStar_matrix}
\usage{
weightedStretchedKendallStar_matrix(x, y, W, stretch = TRUE)
}
\arguments{
\item{x}{numeric vector of non-negative data x}

\item{y}{numeric vector of non-negative data y}

\item{W}{weight matrix, same columns as length x and y}

\item{stretch}{logical default TRUE, whether to stretch based on calculated upper and lower bounds}
}
\value{
\code{vector} of weighted correlations for the sequence of weights given
}
\description{
The weightedStretchedKendallStar_matrix function calculates the stretched weighted Tau*, where where Tau* is described in Pimentel et al (2015) doi: 10.1016/j.spl.2014.09.002, and upper and lower bounds given in Denuit et al (2017) doi: 10.1016/j.spl.2017.03.005
}
\examples{

x = pmax(0,rnorm(100))
y = pmax(0,rnorm(100))
W = weightMatrix(100)
wcor = weightedStretchedKendallStar_matrix(x, y, W)

}
