% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{weightedVariance}
\alias{weightedVariance}
\title{weightedVariance}
\usage{
weightedVariance(x, y = NULL, w)
}
\arguments{
\item{x}{x is a data vector}

\item{y}{default to NULL, if given it is ignored}

\item{w}{weight vector, values should be between 0 and 1}
}
\value{
\code{numeric} weighted variance value for x
}
\description{
the weightedVariance function
}
\examples{

x = rnorm(100)
w = runif(100)
weightedVariance(x,w)

}
