% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCARS.R
\name{weightMatrix}
\alias{weightMatrix}
\title{weightMatrix}
\usage{
weightMatrix(n, type = "triangular", span = NULL, plot = FALSE)
}
\arguments{
\item{n}{the number of samples, same as number of columns of data}

\item{type}{one of "harmonic", "triangular" or "block"}

\item{span}{proportion of samples to include on either side, default is 0.5}

\item{plot}{if TRUE, a heatmap of weight matrix will be generated}
}
\value{
\code{matrix} an n*n matrix is generated corresponding to the weights for each sample
}
\description{
the weightMatrix function
}
\examples{

weightMatrix(100)
weightMatrix(100, plot = TRUE)
weightMatrix(100, type = "triangular", span = 0.1, plot = TRUE)
weightMatrix(100, type = "block", plot = TRUE)
weightMatrix(100, type = "harmonic", plot = TRUE)

}
